Add a test in the feature for sites maps
And in the process uncover what the problem was with viewing the sites
without a query string - it's to do with the string munging we've
done on the GeoJSON URL. We can't use plain routing to assemble the
GeoJSON URL. `sites_path(current_scopes)` will pass through the
scopes correctly, so if `current_scopes` == { with_scores: 1 }, the URL
will be `/sites?with_scores=1`. However, an apparent bug (?) in rails
routing means that if we try `sites_path(current_scopes, format:
:geojson)`, the presence of `format` will mean the scopes are cleared.
We have to generate the URL without specifying the format, then munge
the format in afterwards - crucially, with or without the query string
question mark.